// src/generated/prisma/enums.ts
export const AccountType = {
  ASSET: 'ASSET',
  LIABILITY: 'LIABILITY',
  BANK: 'BANK',
  CAPITAL: 'CAPITAL',
  CASH: 'CASH',
  INCOME: 'INCOME',
  EXPENSE: 'EXPENSE',
  OTHER_EXPENSE: 'OTHER_EXPENSE'
} as const

export type AccountType = (typeof AccountType)[keyof typeof AccountType]